Fixes issue 1768.
Add ilbc and g729 frame per packet options. Could be extended to other codecs on demand.
Doesn't include iLBC mode yet because not sure on how webRTC manage that yet. However fix the fact the mode was not sent in SDP, and treat remote side SDP params for iLBC in webRTC.